Add back call to EVP_CipherFinal_ex in decryption phase.
Few other miscellaneous cleanups. Looks lke we're still tripping over some
undefined behaviour, I see crashes and/or corruption that changes depending
on compiler (gcc/clang) and optimization.

+++ crc32c_test.cc
@@ -21,7 +21,7 @@ int main(int argc, char **argv) {
   for (int i = 0; i < 100; i++) {
     size_t len = size_dist(rndeng);
     buf.resize(len);
-    for (int j = 0; j < len; j++) {
+    for (size_t j = 0; j < len; j++) {
       buf[j] = d_dist(rndeng);
     }
     uint32_t crc_hw = crc32c_hw(buf.data(), len);
